Star At Abim

+++ src/commands/Sıfırlama/isimsil.js
@@ -32,6 +32,6 @@ return;
 }
 message.react(green)
 message.lineReply(`${green} ${member} üyesinin isim geçmişi ${message.author} tarafından \`${moment(Date.now()).format("LLL")}\` tarihinde temizlendi!`)
-await isim.deleteMany({})
+await isim.deleteMany({userID: member.user.id, guildID: message.guild.id})
 }
 };
